aml-dev-tools/automation/requirements.txt
autopep8==1.2.4
backports-abc==0.4
certifi==2017.7.27.1
chardet==3.0.4
flake8==3.4.1
gitdb==0.6.4
GitPython==2.0.2
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
pep8==1.7.0
pycodestyle==2.3.1
pyflakes==1.5.0
requests==2.18.4
six==1.10.0
smmap==0.9.0
urllib3==1.22
